// server/interface/request.interface.ts
export interface SendSmsParams {
  captcha?: string
  captchaId?: string
  mobile: string
}
// ...
export interface RegistersByCodeParams {
  userType: 'user' | 'corp'
  userPhone: string
  customerPassword: string
  invitePhone: string
  smsCode: string
  captchaId: string
  busiCode: string
}